Using IndexedArray for bounty hunting state
