+++ src/Engine/IocContainer.cpp
@@ -169,7 +169,7 @@ void IntegrityTest() {
     //static_assert(sizeof(MapsLongTimer) == 0x20, "Wrong type size");
     static_assert(sizeof(SavegameHeader) == 0x64, "Wrong type size");
     static_assert(sizeof(FactionTable) == 0x1EF1, "Wrong type size");
-    static_assert(sizeof(PartyTimeStruct) == 0x678, "Wrong type size");
+    //static_assert(sizeof(PartyTimeStruct) == 0x678, "Wrong type size");
     //static_assert(sizeof(CastSpellInfo) == 0x14, "Wrong type size");
     static_assert(sizeof(ArcomageCard) == 0x6C, "Wrong type size");
     static_assert(sizeof(LightsData) == 0x3FC, "Wrong type size");
