+++ src/Engine/Objects/Player.cpp
@@ -33,6 +33,7 @@
 #include "GUI/GUIWindow.h"
 #include "GUI/UI/UIGame.h"
 #include "GUI/UI/UIStatusBar.h"
+#include "GUI/UI/UIHouses.h"
 
 #include "Utility/Memory/MemSet.h"
 #include "Utility/IndexedArray.h"
