+++ src/GUI/GUIWindow.cpp
@@ -55,7 +55,7 @@ GUIWindow *pPrimaryWindow;
 
 GUIWindow *pGUIWindow_CurrentMenu;
 GUIWindow *pDialogueWindow;
-GUIWindow *window_SpeakInHouse;
+GUIWindow_House *window_SpeakInHouse;
 GUIWindow_Scroll *pGUIWindow_ScrollWindow; // reading a message scroll
 GUIWindow *ptr_507BC8;  // screen 19 - not used?
 GUIWindow *pGUIWindow_CastTargetedSpell;
@@ -1601,14 +1601,14 @@ void CheckBountyRespawnAndAward() {
     pDialogueWindow->_41D08F_set_keyboard_control_group(1, 1, 0, 2);
     dialog_menu_id = DIALOGUE_OTHER;
     // get new monster for hunting
-    if (pParty->PartyTimes.bountyHunting_next_generation_time[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE] <
+    if (pParty->PartyTimes.bountyHunting_next_generation_time[window_SpeakInHouse->houseId()] <
         pParty->GetPlayingTime()) {
-        pParty->monster_for_hunting_killed[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE] = false;
-        pParty->PartyTimes.bountyHunting_next_generation_time[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE] =
+        pParty->monster_for_hunting_killed[window_SpeakInHouse->houseId()] = false;
+        pParty->PartyTimes.bountyHunting_next_generation_time[window_SpeakInHouse->houseId()] =
             GameTime((int64_t)((double)(0x12750000 * (pParty->uCurrentMonth + 12 * pParty->uCurrentYear - 14015)) * 0.033333335));
         for (;;) {
             rand_monster_id = grng->Random(258) + 1;
-            pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE] = rand_monster_id;
+            pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()] = rand_monster_id;
             if ((uint16_t)rand_monster_id < 0x73u || (uint16_t)rand_monster_id > 0x84u) {
                 if (((uint16_t)rand_monster_id < 0xEBu ||
                     (uint16_t)rand_monster_id > 0xFCu) && ((uint16_t)rand_monster_id < 0x85u ||
@@ -1619,19 +1619,19 @@ void CheckBountyRespawnAndAward() {
             }
         }
     }
-    bountyHunting_monster_id_for_hunting = pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE];
-    if (!pParty->monster_for_hunting_killed[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE]) {
+    bountyHunting_monster_id_for_hunting = pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()];
+    if (!pParty->monster_for_hunting_killed[window_SpeakInHouse->houseId()]) {
         bountyHunting_text = pNPCTopics[351].pText;
-        if (!pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE])
+        if (!pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()])
             bountyHunting_text = pNPCTopics[353].pText;
     } else {  // get prize
-        if (pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE]) {
-            pParty->PartyFindsGold(100 * pMonsterStats->pInfos[(uint16_t)pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE]].uLevel, 0);
+        if (pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()]) {
+            pParty->PartyFindsGold(100 * pMonsterStats->pInfos[(uint16_t)pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()]].uLevel, 0);
             for (uint i = 0; i < 4; ++i)
                 pParty->pPlayers[i].SetVariable(VAR_Award, Award_BountiesCollected);
-            pParty->uNumBountiesCollected += 100 * pMonsterStats->pInfos[pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE]].uLevel;
-            pParty->monster_id_for_hunting[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE] = 0;
-            pParty->monster_for_hunting_killed[window_SpeakInHouse->wData.val - HOUSE_TOWNHALL_HARMONDALE] = false;
+            pParty->uNumBountiesCollected += 100 * pMonsterStats->pInfos[pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()]].uLevel;
+            pParty->monster_id_for_hunting[window_SpeakInHouse->houseId()] = 0;
+            pParty->monster_for_hunting_killed[window_SpeakInHouse->houseId()] = false;
         }
         bountyHunting_text = pNPCTopics[352].pText;
     }
